test(payments): close the audit's remaining test-side gaps
- The ackBatch contract tests opened with `if (port.ackBatch === undefined)
  return`, so they would have passed green against a port implementing nothing.
  They now assert the method exists, matching how the deliverBatch case has
  always done it.
- hookedDelivery in the facade harness hand-forwards a fixed list of methods and
  had silently dropped deliverBatch and onWake; ackBatch was the third. Optional
  methods are forwarded when the inner port has them, so facade tests exercise
  the path production takes instead of a fallback.
- coalesced's tests moved into async.test.ts, beside SingleFlight and
  SerialChain. That file's own header records this repo collapsing seven
  hand-rolled copies into async.ts once already.

400 payments tests pass (3 more than before: the facade suite now runs the batch
path); 2019 across the SDK.
